chore: improve test stability
* SalteAuthProfile and SalteAuthUtilities are no longer singletons
* Added a $navigate function for coverage purposes
* Split utility tests into separate files

closes #99 and closes #100
