+++ src/salte-auth.js
@@ -93,7 +93,7 @@ class SalteAuth {
       if (error) {
         this.profile.$clear();
       } else {
-        location.href = this.profile.$redirectUrl;
+        this.$utilities.$navigate(this.profile.$redirectUrl);
         this.profile.$redirectUrl = undefined;
       }
       this.$config.redirectLoginCallback(error);
@@ -307,7 +307,7 @@ class SalteAuth {
 
     this.profile.$clear();
     this.profile.$redirectUrl = this.profile.$redirectUrl || location.href;
-    location.href = this.$loginUrl;
+    this.$utilities.$navigate(this.$loginUrl);
 
     return this.$promises.login;
   }
@@ -367,7 +367,7 @@ class SalteAuth {
    */
   logoutWithRedirect() {
     this.profile.$clear();
-    location.href = this.$deauthorizeUrl;
+    this.$utilities.$navigate(this.$deauthorizeUrl);
   }
 
   /**
@@ -427,3 +427,4 @@ class SalteAuth {
 
 set(window, 'salte.SalteAuth', get(window, 'salte.SalteAuth', SalteAuth));
 export { SalteAuth };
+export default SalteAuth;
