+++ src/salte-auth.providers.js
@@ -41,3 +41,4 @@ class Providers {
 };
 
 export { Providers };
+export default Providers;
