+++ src/salte-auth.utilities.js
@@ -1,6 +1,4 @@
 import assign from 'lodash/assign';
-import get from 'lodash/get';
-import set from 'lodash/set';
 
 /**
  * Basic utilities to support the authentication flow
@@ -10,10 +8,6 @@ class SalteAuthUtilities {
    * Wraps all XHR and Fetch (if available) requests to allow promise interceptors
    */
   constructor() {
-    if (window.salte.SalteAuthUtilities.$instance) {
-      return window.salte.SalteAuthUtilities.$instance;
-    }
-
     /** @ignore */
     this.$interceptors = {
       fetch: [],
@@ -61,7 +55,6 @@ class SalteAuthUtilities {
         };
       })(fetch);
     }
-    window.salte.SalteAuthUtilities.$instance = this;
   }
 
   /**
@@ -266,7 +259,17 @@ class SalteAuthUtilities {
     }
     return null;
   }
+
+  /**
+   * Navigates to the url provided.
+   * @param {String} url the url to navigate to
+   * @private
+   */
+  /* istanbul ignore next */
+  $navigate(url) {
+    location.href = url;
+  }
 }
 
-set(window, 'salte.SalteAuthUtilities', get(window, 'salte.SalteAuthUtilities', SalteAuthUtilities));
 export { SalteAuthUtilities };
+export default SalteAuthUtilities;
