+++ tests/salte-auth/providers/auth0.spec.js
@@ -1,16 +1,14 @@
 import { expect } from 'chai';
 
-import { SalteAuth } from '../../../src/salte-auth.js';
+import SalteAuthUtilities from '../../../src/salte-auth.utilities.js';
 import auth0 from '../../../src/providers/auth0.js';
 
-const auth = new SalteAuth({
-  provider: 'auth0'
-});
-
 describe('auth0', () => {
+  const utilities = new SalteAuthUtilities();
+
   describe('function(deauthorizeUrl)', () => {
     it('should create a logout url', () => {
-      const url = auth0.deauthorizeUrl.call(auth, {
+      const url = auth0.deauthorizeUrl.call({ $utilities: utilities }, {
         providerUrl: 'https://api.salte.io',
         redirectUrl: `${location.protocol}//${location.host}`,
         clientId: '33333333-3333-4333-b333-333333333333'
