+++ tests/salte-auth/providers/cognito.spec.js
@@ -1,16 +1,14 @@
 import { expect } from 'chai';
 
-import { SalteAuth } from '../../../src/salte-auth.js';
+import SalteAuthUtilities from '../../../src/salte-auth.utilities.js';
 import cognito from '../../../src/providers/cognito.js';
 
-const auth = new SalteAuth({
-  provider: 'cognito'
-});
-
 describe('cognito', () => {
+  const utilities = new SalteAuthUtilities();
+
   describe('function(authorizeEndpoint)', () => {
     it('should create a authorize endpoint', () => {
-      expect(cognito.authorizeEndpoint.call(auth, {
+      expect(cognito.authorizeEndpoint.call({ $utilities: utilities }, {
         providerUrl: 'https://mydomain.auth.us-east-1.amazoncognito.com'
       })).to.equal(`https://mydomain.auth.us-east-1.amazoncognito.com/oauth2/authorize`);
     });
@@ -18,7 +16,7 @@ describe('cognito', () => {
 
   describe('function(deauthorizeUrl)', () => {
     it('should create a logout url', () => {
-      const url = cognito.deauthorizeUrl.call(auth, {
+      const url = cognito.deauthorizeUrl.call({ $utilities: utilities }, {
         providerUrl: 'https://mydomain.auth.us-east-1.amazoncognito.com',
         redirectUrl: `${location.protocol}//${location.host}`,
         clientId: '33333333-3333-4333-b333-333333333333'
